Move interval for countdown into countdown component
